add compact home audiogram with background and a little border
